requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langchain>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
transformers>=4.0.0
sentence-transformers>=2.0.0
torch>=2.0.0
